smart teleop test code & new arm and claw test code
