+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_TeleOp/Old_Arm_Claw/_2023_11_06_07_Arm_Claw_V7.java
@@ -1,7 +1,6 @@
-package org.firstinspires.ftc.teamcode.Developing_Code.Old_Arm_Claw;
+package org.firstinspires.ftc.teamcode._TeleOp.Old_Arm_Claw;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
@@ -10,9 +9,8 @@ import com.qualcomm.robotcore.hardware.Servo;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 
-@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V5 Arm Claw")
-@Disabled
-public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
+@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V7 Arm Claw + Telemetry")
+public class _2023_11_06_07_Arm_Claw_V7 extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
         DcMotor frontLeftMotor = hardwareMap.dcMotor.get("motorFL");
@@ -96,10 +94,10 @@ public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
             double frontRightPower = ((rotY - rotX - rx) / denominator);
             double backRightPower = ((rotY + rotX - rx) / denominator);
 
-            frontLeftMotor.setPower(frontLeftPower * 0.65);
-            backLeftMotor.setPower(backLeftPower * 0.65);
-            frontRightMotor.setPower(frontRightPower * 0.65);
-            backRightMotor.setPower(backRightPower * 0.65);
+            frontLeftMotor.setPower(frontLeftPower * 0.8);
+            backLeftMotor.setPower(backLeftPower * 0.8);
+            frontRightMotor.setPower(frontRightPower * 0.8);
+            backRightMotor.setPower(backRightPower * 0.8);
 
             double liftPower = arm.getPower();
 
@@ -116,14 +114,14 @@ public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
             }
 
             if (gamepad1.right_bumper) {
-                for (int liftPos = 0; liftPos <= 500; liftPos = liftPos + 5) {
+                axle.setPosition(0.8);
+                for (int liftPos = 0; liftPos <= 500; liftPos = liftPos + 4) {
                     liftTargetPosition = liftPos;
                     arm.setTargetPosition(liftTargetPosition);
                     arm.setPower(0.7);
                     arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-                    sleep(15);
+                    sleep(12);
                 }
-
                 clawLeft.setPosition(openClaw);
                 clawRight.setPosition(openClaw);
                 
@@ -133,7 +131,7 @@ public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
             if (gamepad1.left_bumper) {
                 liftTargetPosition = 10;
                 arm.setTargetPosition(liftTargetPosition);
-                arm.setPower(0.1);
+                arm.setPower(0.07);
                 arm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
                 
                 axle.setPosition(0.8);
@@ -163,6 +161,35 @@ public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
             if (gamepad1.dpad_down) {
                 axle.setPosition(0.8);
             }
+            if (gamepad1.dpad_left) {
+                axle.setPosition(0.8);
+                clawLeft.setPosition(openClaw);
+                clawRight.setPosition(openClaw);
+                sleep(200);
+
+                backLeftMotor.setPower(0.3);
+                backRightMotor.setPower(0.3);
+                frontLeftMotor.setPower(0.3);
+                frontRightMotor.setPower(0.3);
+                sleep(200);
+                backLeftMotor.setPower(0);
+                backRightMotor.setPower(0);
+                frontLeftMotor.setPower(0);
+                frontRightMotor.setPower(0);
+
+                clawLeft.setPosition(closeClaw);
+                clawRight.setPosition(closeClaw);
+
+                backLeftMotor.setPower(-0.3);
+                backRightMotor.setPower(-0.3);
+                frontLeftMotor.setPower(-0.3);
+                frontRightMotor.setPower(-0.3);
+                sleep(200);
+                backLeftMotor.setPower(0);
+                backRightMotor.setPower(0);
+                frontLeftMotor.setPower(0);
+                frontRightMotor.setPower(0);
+            }
 
             telemetry.addData("Claw Left Target: ", clawLeftTarget);
             telemetry.addData("Claw Right Target: ", closeClaw);
@@ -176,6 +203,8 @@ public class _2023_10_27_05_Arm_Claw_V5 extends LinearOpMode {
             }
 
             // ADDED CODE - sends info about current servo position to driver station
+            telemetry.addData("Bot Heading: ", botHeading);
+
             telemetry.addData("Drone Servo Position: ", droneServoPosition);
 
             telemetry.addData("Claw Left Position: ", clawLeft.getPosition());
