+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/_TeleOp/Virtual_4Bar/_2023_11_25_01_Virtual_4Bar_V1.java
@@ -1,26 +1,36 @@
-package org.firstinspires.ftc.teamcode.Developing_Code.Virtual_4Bar;
+package org.firstinspires.ftc.teamcode._TeleOp.Virtual_4Bar;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Servo;
 
-@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V1 Virtual 4Bar No Drivebase")
-public class _2023_11_25_01_Virtual_4Bar_No_DB_V1 extends LinearOpMode {
+@com.qualcomm.robotcore.eventloop.opmode.TeleOp(name = "V1 Virtual 4Bar")
+public class _2023_11_25_01_Virtual_4Bar_V1 extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
+        DcMotor frontLeftMotor = hardwareMap.dcMotor.get("motorFL");
+        DcMotor backLeftMotor = hardwareMap.dcMotor.get("motorBL");
+        DcMotor frontRightMotor = hardwareMap.dcMotor.get("motorFR");
+        DcMotor backRightMotor = hardwareMap.dcMotor.get("motorBR");
+
         DcMotor arm = hardwareMap.dcMotor.get("arm");
-        //Servo clawTop = hardwareMap.servo.get("clawTop");
+        Servo clawTop = hardwareMap.servo.get("clawTop");
         Servo clawBottom = hardwareMap.servo.get("clawBottom");
 
+        frontRightMotor.setDirection(DcMotorSimple.Direction.FORWARD);
+        backRightMotor.setDirection(DcMotorSimple.Direction.FORWARD);
+        frontLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+        backLeftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+
         arm.setDirection(DcMotorSimple.Direction.REVERSE);
-        //clawTop.setDirection(Servo.Direction.FORWARD);
+        clawTop.setDirection(Servo.Direction.FORWARD);
         clawBottom.setDirection(Servo.Direction.REVERSE);
 
         arm.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         arm.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        //clawTop.scaleRange(0, 1);
+        clawTop.scaleRange(0, 1);
         clawBottom.scaleRange(0, 1);
 
         //clawBottom.setPosition(0);
@@ -33,10 +43,33 @@ public class _2023_11_25_01_Virtual_4Bar_No_DB_V1 extends LinearOpMode {
 
         waitForStart();
 
-        //clawTop.setPosition(openClaw);
+        clawTop.setPosition(openClaw);
         clawBottom.setPosition(openClaw);
 
         while (opModeIsActive()) {
+            //gamepad 1 - drive base
+            double y = -gamepad1.left_stick_y; // Remember, Y stick value is reversed
+            double x = gamepad1.left_stick_x;
+            double rx = gamepad1.right_stick_x;
+
+            x = x * 1.1;  // Counteract imperfect strafing
+
+            // Denominator is the largest motor power (absolute value) or 1
+            // This ensures all the powers maintain the same ratio,
+            // but only if at least one is out of the range [-1, 1]
+            double denominator = Math.max(Math.abs(y) + Math.abs(x) + Math.abs(rx), 1);
+            double frontLeftPower = Math.round(((y + x + rx) / denominator));
+            double backLeftPower = Math.round(((y - x + rx) / denominator));
+            double frontRightPower = Math.round(((y - x - rx) / denominator));
+            double backRightPower = Math.round(((y + x - rx) / denominator));
+
+            frontLeftMotor.setPower(frontLeftPower * 0.8);
+            backLeftMotor.setPower(backLeftPower * 0.8);
+            frontRightMotor.setPower(frontRightPower * 0.8);
+            backRightMotor.setPower(backRightPower * 0.8);
+
+            double liftPower = arm.getPower();
+
             if (gamepad1.right_trigger > 0) {
                 liftTargetPosition += 2.5;
             } else if (gamepad1.left_trigger > 0) {
@@ -114,6 +147,7 @@ public class _2023_11_25_01_Virtual_4Bar_No_DB_V1 extends LinearOpMode {
             telemetry.addData("Claw Bottom Position: ", clawBottom.getPosition());
 
             telemetry.addData("Arm position: ", arm.getCurrentPosition());
+            telemetry.addData("Arm power: ", liftPower);
             telemetry.addData("Arm Target Position Requested: ", liftTargetPosition);
             telemetry.addData("Arm Actual Target Position: ", arm.getTargetPosition());
